NodeSet is internal class of MongoDBSystem, removed ConnectionManager
