+++ driver/src/main/scala/core/nodeset.scala
@@ -41,78 +41,6 @@ package object utils {
   }
 }
 
- case class NodeSet(connectionManager: ActorRef)
-   extends Actor with ActorLogging {
-
-   var initialAuthenticates: Seq[Authenticate] = Seq.empty
-   var connectionsPerNode: Int = 10
-   var existingHosts : Set[String] = Set.empty
-   var connectingNodes: Vector[ActorRef] = Vector.empty
-   var connectedNodes: Vector[ActorRef] = Vector.empty
-   var version: Option[Long] = None
-   var replyTo: ActorRef = null
-
-  override def receive: Receive = {
-    case NodeSet.ConnectAll(hosts, auth, count) => {
-      log.info("Connection to initial nodes")
-      replyTo = sender()
-      this.connectionsPerNode = count
-      this.initialAuthenticates = auth
-      existingHosts = hosts.toSet
-      connectingNodes = hosts.map(address => {
-        val node = context.actorOf(Props(classOf[Node], address, initialAuthenticates, connectionsPerNode))
-        node ! Node.Connect
-        node
-      }).toVector
-    }
-    case Node.Connected(connections, metadata) => {
-      log.info("node connected metadata {}", metadata)
-      connectingNodes = connectingNodes diff List(sender())
-      connectedNodes = sender() +: connectedNodes
-      connections.foreach(p => connectionManager ! Add(p._1, p._2))
-
-      if(connections.exists(p => p._2.isPrimary || p._2.isMongos))
-        replyTo ! metadata
-    }
-    case Node.DiscoveredNodes(hosts) => {
-      log.info("nodes descovered")
-      val discovered = hosts.filter(!existingHosts.contains(_))
-      existingHosts = discovered ++: existingHosts
-      connectingNodes = discovered.map(address => {
-        val node = context.actorOf(Props(classOf[Node], address, initialAuthenticates, connectionsPerNode))
-        node ! Node.Connect
-        node
-      }) ++: connectingNodes
-    }
-    case Close => {
-      connectedNodes.foreach(_ ! Close)
-      context.become(closing)
-    }
-  }
-
-   private def closing: Receive = {
-     case Node.Connected(connections, metadata) => {
-       log.info("node connected metadata {} but nodeSet in closing state", metadata)
-       connectingNodes = connectingNodes diff List(sender())
-       connectedNodes = sender() +: connectedNodes
-       sender() ! Close
-     }
-     case Closed => {
-      connectedNodes = connectedNodes diff List(sender())
-       if(connectedNodes.isEmpty && connectingNodes.isEmpty) {
-         self ! PoisonPill
-         log.debug("all connections to a nodeset are closed")
-       }
-     }
-     case Node.DiscoveredNodes(hosts) => {
-       log.info("nodes descovered but nodeSet in closing state")
-     }
-   }
-}
-
-object  NodeSet {
-  case class ConnectAll(hosts: Seq[String], initialAuthenticates: Seq[Authenticate], connectionsPerNode: Int)
-}
 
 
 case class ProtocolMetadata(
